src/modules/cloud-height/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clouds-decoded-cloud-height"
version = "0.1.0"
description = "Cloud Height Retrieval Module"
requires-python = ">=3.10"
dependencies = [
    "clouds-decoded-core",
    "numpy",
    "scipy",
    "rasterio",
    "pyproj",
    "rtree",
    "tqdm",
    "pyyaml",
    "matplotlib"
]

[tool.hatch.build.targets.wheel.force-include]
"." = "clouds_decoded/modules/cloud_height"
